finish layout (witch a catch)
I put the led matrix footprint on the wrong side of the board...
This was fixed by moving the pin headers which however now results in
the controller IC being under the led matrix module. Not what I wanted
but not easy to fix. I wanted to use headers to mount the module anyways
to there should be no interference. Try to use some short female pin
headers (which apparently is a thing).